create useFile hook
